wmcs infra-tracing: optimize Loki indexing
Why:
When sending logs to loki, the stream labels are indexed for more
efficient querying while the per-line labels are not.

What:
Change the way the data is pushed to Loki to split it into multiple
streams, each one with a set of unique labels.
Use the same set of labels during the buffering of log lines using
immutable datacalss instances to represent the set of labels.

Increase the default buffering of log lines to 300 based on actual
traffic in toolforge.

Notify the systemd unit if the Python script changes too, not only the
INI config file.

Hosts: auto
Bug: T399313
Change-Id: I479251322cd6b8a371acea763e6fe2ffcec8471d