Redirect mpic-next.w.o to test-kitchen-next.w.o
This will allow us to automatically redirect human visitors of mpic-next.w.o
to the renamed test-kitchen-next.w.o.

Once this is done, we should be able to login via CAS with a callback
URL set to https://test-kitchen-next.w.o/<path>

If that works well, we'll then proceed to migrating mpic production.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T407805
Change-Id: I7d7dba59add95562e45a55493c21751e9a50688d