admin/releases: deprecate shell user group releasers-mwcli
Deprecate the shell user group that gives access to upload
releases on https://releases.wikimedia.org/mwcli/

There is only one initial release here and later releases
are on https://gitlab.wikimedia.org/repos/releng/cli/-/releases

So there does not seem to be a manual upload of releases anymore.

In discussion.

Possibly we should link the old place to the new place
or automate uploads?

Change-Id: I0e54cff242924e5dd2055e9d3c58079ca676f499