openstack: puppet: Do not commit empty role fiels
In the set_roles endpoint, instead of committing the given data as is
(which can be an empty array), use logic similar to the newer
update_prefix_by_id endpoint which round-trips it back to YAML in the
code. This ensures consistent formatting and also means we don't commit
empty files.

Change-Id: I19fcfd966295c3d0ea9439bc690734096a6a6964