P:toolforge: static: Publish worker IPs as a JSON file
Publish a dump of the Toolforge K8s worker IPs so that they can be used
to tweak the rate limiting rules on the wikis.

This patch also creates wmflib::ip2cidr as a generic function since this
is a relatively common pattern in this repo, I'll send some follow-ups
converting uses to the new function.

Bug: T411610
Change-Id: I076f5f6d72a748831f5dab7394a86cb66a6a6964