Add more approvers for analytics-platform-eng-admins group
This group is used to allow members to sudo to the
analytics-platform-eng user, in order to perform hive and hadoop related
operations on behalf of that system user.

We need to add a new user to the group, but the only approver currently
listed is Andrew Otto, who is on new parent leave until early 2026.

We wish to add more approvers to the group, prior to adding a new
group member.

Bug: T411648
Change-Id: I9e3c0c95d031d51a9b486f442a043fb9d84c2b29