Add a growthbook system user and grant it access to private data
We need to make a new system user for growthbook, since this will be
querying presto, but is currently unable to use user impersonation.

Adding the 'growthbook' system user to the 'analytics-privatedata-users'
group will allow growthbook to read the event tables where it sources
its data for experiment analysis.

Hosts: P:hadoop::common
Bug: T406593
Change-Id: I9dd6bcd65d31652ffc83c24061b6cb82fad5538a