varnish: remove ancient Noise rule from text-frontend VCL
This kind of special rule looks like it should be replaced by
a requestctl action nowadays.

Or be completely removed as outdated?

I tracked it back to 2015 in 225f38ada6a but even that refers
to it as being moved from somewhere else.

Depending on the reaction to this I would also question
the next 2 rules following this one.

Change-Id: I5e0eed5b2af025baac405d19143b1873827d84b7