P:cache::haproxy mark requests from WMCS as trusted
Requests coming from WMCS should always be marked as trusted.
Currently this is always the case, as their IPs are in the private
ranges. The IP range feed is managed by WMCS and may contain non-
private IPs in the future, checking the cloud provider provenance
ensures that these IPs will also be marked as trusted in the future.

Bug: T411503
Change-Id: Ia9720b98fb743e2c5ad1b37431468cb218fb6d83