Allow wdqs10[28-32] to mount the NFS dumps directories
We are using several wdqs hosts to begin testing qlever and virtuoso.

As such, it would be useful if they could mount the xmldatadumps volumes
from the dumps distribution servers, so that they can access the recent
wikidata entity dumps.

This commit adds the five hosts in the wdqs::test::alternative role to
the list of hosts permitted to mount the volume.

Hosts: O:dumps::distribution::server
Bug: T412351
Change-Id: I803fc19a3ca6719aaf4028b8f7e6e7005597a3ee