DNS discovery: split responses to magru servers based on rack
This should have the effect to change which core site (eqiad/codfw)
DNS record is returned to hosts in magru based on which rack the host
making the request is in.

For active/passive services this won't make a difference.  For
active-active services (like Swift) it will mean that requests
from magru will be spread across both core sites, which will
lead to a balance of traffic on our backhaul links.

NOT to be merged unless we decide this is something we want/need to
do, but better to have the patch ready and reviewed.

Bug: T411617
Change-Id: I6ec1f2c9876c93004a3aee9b749a7798670f37d4