Trixie d-i preseed file: increase link_wait_timeout
As described on task we are having problems reimaging some hosts,
where the debian installer fails to indentify the correct link to
apply IP configuration passed to it on.

The intent of this is to give the link more time to come up, to see
if it can identify the correct one if given a little more time. It
seems unlikely but worth a test, if it does help we can decide on
the right longer-term option.

Bug: T412807
Change-Id: I464be7c8be944dd2606f408440d937d9e89ed646