autoinstall: set timeout for network link detection for uefi mode
Moritz pointed out that setting this network option in the preseed
file, which the host downloads once the network is up, isn't going
to be any use to set network options required to bring the network
up in the first place!

Removing the option from preseed here, and adding it to the options
for ipxe uefi boot.  If it works we can also adjust the files for
tftpboot and other modes.

Bug: T412807
Change-Id: Ic5765775f3cedf53a1a4c2e2bda1ecb9e5eb6255